LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/javascript-url-security-check-failure.sub-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://www.web-platform.test:8800". Protocols, domains, and ports must match.
CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://www.web-platform.test:8800". Protocols, domains, and ports must match.
CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://web-platform.test:8800". The frame being accessed set "document.domain" to "web-platform.test", but the frame requesting access did not. Both must set "document.domain" to the same value to allow access.
CONSOLE MESSAGE: Blocked a frame with origin "http://web-platform.test:8800" from accessing a frame with origin "http://web-platform.test:8800". The frame being accessed set "document.domain" to "web-platform.test", but the frame requesting access did not. Both must set "document.domain" to the same value to allow access.

PASS cross-origin, setting src
PASS cross-origin, setting location.href
PASS cross-origin-domain but same-origin, setting src
PASS cross-origin-domain but same-origin, setting location.href
